- created a maven multi module project - refactored the filestructure to reflect the maven multi module - updated versions
